feat:configurando as models de imagem e usuario
