+++ app/models/__all__models.py
@@ -1,7 +1,8 @@
 
 # Adiciona todos os imports da models aqui para fazer as migrations no banco de dados
 from .product_models import ProductModel
-from .usuario_models import UsuarioModel,ImagensModel
+from .usuario_models import UsuarioModel
 from .usuario_google_models import UsuarioGoogleModel
 from .vendedor_models import VendedorModel
+from .image_model import ImagensModel
