+++ app/models/usuario_models.py
@@ -4,16 +4,6 @@ from sqlalchemy.orm import relationship
 
 from app.config.configs import settings
 
-
-class ImagensModel(settings.DBBaseModel):
-    __tablename__ = "imagens"
-    id: int = Column(Integer, primary_key=True, autoincrement=True)
-    url: str = Column(String(256))
-    nome: str = Column(String(256))
-    id_usuario: int = Column(Integer, ForeignKey("usuario.id"))
-
-    usuario = relationship("UsuarioModel", back_populates="imagem")
-
 class UsuarioModel(settings.DBBaseModel):
     __tablename__ = "usuario"
 
@@ -23,6 +13,6 @@ class UsuarioModel(settings.DBBaseModel):
     sobrenome: str = Column(String(256))
     celular: str = Column(String(13))
     senha: str = Column(String(256))
-    imagem = relationship("ImagensModel", back_populates="usuario")
+    imagem = relationship("ImagensModel", uselist=False,backref="usuario")
     vendedor = relationship("VendedorModel",  uselist=False, backref="usuario")
