+++ app/routes/api_config.py
@@ -2,7 +2,7 @@
 from fastapi import APIRouter
 
 
-from .v1_endpoint import usuario_routes, produto_routes, vendedor_routes
+from .v1_endpoint import usuario_routes, produto_routes, vendedor_routes,image_routes
 
 
 
@@ -12,3 +12,4 @@ api_router = APIRouter()
 api_router.include_router(usuario_routes.router, prefix='/usuarios', tags=['usuarios'])
 api_router.include_router(vendedor_routes.router, prefix='/vendedor', tags=['vendedor'])
 api_router.include_router(produto_routes.router, prefix='/produtos', tags=['produtos'])
+api_router.include_router(image_routes.router, prefix='/imagens', tags=['imagens'])
