+++ app/schemas/image_schemas.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel, EmailStr, HttpUrl
+
+class Imagem(BaseModel):
+    url: HttpUrl
+    nome: str
